fill-station/client/requirements.txt
streamlit==1.40.0
websocket-client==1.8.0
pandas==2.2.3
plotly==6.0.0
websockets==12.0  # For mock server
